// src/interfaces/CreditCardStyle.interface.ts
export interface ICreditCardStyle{
    backgroundColor: string;
    backgroundImage: string;
    logoImage: string;
    margin: string;
    chip: string;
    opacity?: string;
    widthLogo?: string;
  
    // American Express:
    marginRow?: string;
    marginChip?: string;	
    fontColor?: string;
    fontWeight?: string;
    textShadow?: string;
  }